move sever code from 'internal' to 'pkg'
