+++ internal/versionbump.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"github.com/ptgoetz/go-versionbump/internal/config"
 	"github.com/ptgoetz/go-versionbump/internal/git"
-	"github.com/ptgoetz/go-versionbump/internal/semver"
 	"github.com/ptgoetz/go-versionbump/internal/utils"
 	vbu "github.com/ptgoetz/go-versionbump/internal/utils"
+	"github.com/ptgoetz/go-versionbump/pkg/semver"
 	"gopkg.in/yaml.v2"
 	"os"
 	"path"
